final_bot-main/requirements.txt
python-telegram-bot[job-queue,webhooks]==22.7
pymongo[srv]==4.6.1
dnspython==2.4.2
